+++ src/Control/Arrow/Machine/Misc/Exception.hs
@@ -1,7 +1,7 @@
 
 
 module
-    Control.Arrow.Machine.Exception
+    Control.Arrow.Machine.Misc.Exception
       (
         catch,
         handle,
@@ -13,38 +13,15 @@ module
        )
 where
 
-import qualified Control.Monad.Trans.Free as F
-import qualified Control.Monad.Trans.Free.Church as F
-
-import Data.Functor ((<$>))
 
 import Control.Arrow.Machine.Types
-import Control.Arrow.Machine.Plan.Internal
-
 
 
 catch :: Monad m =>
     PlanT i o m a -> PlanT i o m a -> PlanT i o m a
 
-catch pl cont = 
-    F.toFT $ catch' (F.fromFT pl) (F.fromFT cont)
+catch = catchP
 
-catch' ::
-    Monad m =>
-    F.FreeT (PlanF t o) m a ->
-    F.FreeT (PlanF t o) m a ->
-    F.FreeT (PlanF t o) m a
-
-catch' (F.FreeT mf) cont@(F.FreeT mcont) = 
-    F.FreeT $ mf >>= go
-  where
-    go (F.Pure a) = return $ F.Pure a
-    go (F.Free StopPF) = mcont
-    go (F.Free (AwaitPF f ff)) = 
-        return $ F.Free $ 
-        AwaitPF (\i -> f i `catch'` cont) (ff `catch'` cont)
-    go (F.Free fft) = 
-        return $ F.Free $ (`catch'` cont) <$> fft
 
 handle :: Monad m =>
     PlanT i o m a -> PlanT i o m a -> PlanT i o m a